api/requirements_minimal.txt
# Minimal requirements for emergency deployment
quart==0.19.9
requests==2.32.3
python-dotenv==1.0.1
azure-communication-callautomation==1.2.0
azure-core==1.32.0
azure-identity==1.15.0
openai==1.57.2
redis==5.2.1
